+++ Dockerfile
@@ -2,10 +2,10 @@ FROM golang:1.18 as builder
 
 WORKDIR /go/src
 
-COPY ../../go.mod ./
+COPY go.mod ./
 RUN go mod download
 
-COPY ../../cmd/identity/main.go ./
+COPY ./*.go ./
 
 ARG CGO_ENABLED=0
 ARG GOOS=linux
