+++ envs/lint/.golangci.yaml
