+++ envs/linux/init.sh
@@ -30,3 +30,4 @@ echo 'export PATH=$PATH:$GOPATH/bin' >> .bashrc
 
 # golangci-lintのインストール
 curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.46.2
+
