+++ envs/linux/lint.sh
@@ -2,9 +2,9 @@
 SCRIPT_DIR=$(dirname "$0")
 cd "$SCRIPT_DIR"
 
-cd ../../../
+cd ../../
 
 GOOS="linux"
 GOARCH="amd64"
 
-golangci-lint run -c ./configs/.golangci.yaml
+golangci-lint run -c ./envs/lint/.golangci.yaml
