+++ envs/macos/init.sh
