+++ envs/macos/lint.sh
