+++ envs/macos/test.sh
