+++ envs/windows/init.bat
