+++ envs/windows/lint.bat
