+++ envs/windows/test.bat
